Jsonifier Release v0.9.91
* Removed store and storeu functions from being class-contained to free-standing.
* Removed unnecessary branch statements from serialize_impl functions.
* Improved some of the preprocessor macros for setting architecture.
* Replaced a function that uses intrinsics with a non-intrinsic (but more rapid) version.
* Renamed parsestring to parseString,.
* Fixed a heap-buffer-overflow issue on two of the comparison functions.
* Added more implementation to the raw_json_data class.
* Added a dedicated concepts namespace.
* Added some string-to-number functions.
* Split the parseNumber function into multiple specializations to improve performance.
* Refactored the jsonifier_object_t and jsonifier_array_t related functions.
* Refactored the AVX-types into a statically-operating class instead of a containing one, to improve performance.
* Renamed a bunch of things.
* Added the structural_index_vector and buffer_string classes for parsing performance.
* Improved error output - for debugging the indices.
* Implemented new serialization logic for strings.
* New subtraction logic.
* Reorganized some of the headers.

+++ Include/jsonifier/Allocator.hpp
@@ -33,8 +33,8 @@ namespace jsonifier_internal {
 		return remainder == 0 ? num : num + (multiple - remainder);
 	}
 
-	template<uint64_t multiple> jsonifier_inline uint64_t roundDownToMultiple(uint64_t value) {
-		return value >= 0 ? (value / multiple) * multiple : ((value - multiple + 1) / multiple) * multiple;
+	template<int64_t multiple> jsonifier_inline uint64_t roundDownToMultiple(int64_t value) {
+		return static_cast<uint64_t>(value >= 0 ? (value / multiple) * multiple : ((value - multiple + 1) / multiple) * multiple);
 	}
 
 	template<typename value_type_new> class aligned_allocator : public std::pmr::polymorphic_allocator<value_type_new> {
